+++ src/Application.php
@@ -2,6 +2,8 @@
 
 namespace Parable\Console;
 
+use Throwable;
+
 class Application
 {
     /**
@@ -53,13 +55,16 @@ class Application
         $this->input     = $input;
         $this->parameter = $parameter;
 
-        set_exception_handler(function (Exception $e) {
+        $throwableClosure = function (Throwable $e) {
             $this->output->writeErrorBlock([$e->getMessage()]);
 
             if ($this->activeCommand) {
                 $this->output->writeln('<yellow>Usage</yellow>: ' . $this->activeCommand->getUsage());
             }
-        });
+        };
+
+        set_exception_handler($throwableClosure);
+        set_error_handler($throwableClosure);
     }
 
     public function setName(string $name): void